refactor: finish fetchAction work & move to api folder
